requirements.txt
sanic>=23.12.1
sanic-ext>=23.12.0
websockets>=12.0
transformers>=4.39.2
requests_futures>=1.0.0
numpy>=1.25.1
scikit-image>=0.21.0
opencv-python>=4.9.0.80
accelerate>=0.22.0
torch>=2.0.1
torchvision>=0.15.2
boto3>=1.28.43
botocore>=1.31.43
pytest>=7.4.2
pytest-cov>=4.1.0
xtarfile[zstd]>=0.1.0
bitsandbytes>=0.41.1
compel>=2.0.2